feat: split directives into separate modules
