+++ source/connect-array/connect-array.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+
+import { ConnectArray } from './connect-array';
+
+const declarations = [ConnectArray];
+
+@NgModule({
+  declarations: [...declarations],
+  exports: [...declarations],
+})
+export class NgReduxFormConnectArrayModule {}
