+++ source/connect-array/index.ts
@@ -1,0 +1,2 @@
+export * from './connect-array.module';
+export * from './connect-array';
