+++ source/index.ts
@@ -3,8 +3,6 @@ export * from './form-reducer';
 export * from './form-exception';
 export * from './form-store';
 export * from './configure';
-export * from './connect-base';
-export * from './connect-reactive';
 export * from './connect';
 export * from './connect-array';
 export * from './module';
