+++ source/module.ts
@@ -1,11 +1,9 @@
 import {NgModule} from '@angular/core';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-
 import {NgRedux} from '@angular-redux/store';
 
-import {ReactiveConnect} from './connect-reactive';
-import {Connect} from './connect';
-import {ConnectArray} from './connect-array';
+import {NgReduxFormConnectModule} from './connect';
+import {NgReduxFormConnectArrayModule} from './connect-array';
 import {FormStore} from './form-store';
 
 export function formStoreFactory(ngRedux: NgRedux<any>) {
@@ -16,16 +14,12 @@ export function formStoreFactory(ngRedux: NgRedux<any>) {
   imports: [
     FormsModule,
     ReactiveFormsModule,
-  ],
-  declarations: [
-    Connect,
-    ReactiveConnect,
-    ConnectArray,
+    NgReduxFormConnectModule,
+    NgReduxFormConnectArrayModule,
   ],
   exports: [
-    Connect,
-    ReactiveConnect,
-    ConnectArray,
+    NgReduxFormConnectModule,
+    NgReduxFormConnectArrayModule
   ],
   providers: [
     {
@@ -33,6 +27,6 @@ export function formStoreFactory(ngRedux: NgRedux<any>) {
       useFactory: formStoreFactory,
       deps: [NgRedux],
     },
-  ]
+  ],
 })
 export class NgReduxFormModule {}
